connectField: Call "mapPropsToField" after "defaultFieldRecord"
