+++ src/Fields/Checkbox.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import createField from '../createField';
+import connectField from '../connectField';
 
 function Checkbox({ fieldProps }) {
   return (<input { ...fieldProps } />);
@@ -7,12 +7,12 @@ function Checkbox({ fieldProps }) {
 
 Checkbox.displayName = 'Checkbox';
 
-export default createField({
+export default connectField({
   valuePropName: 'checked',
-  mapPropsToField: props => ({
-    ...props,
+  mapPropsToField: ({ props: { checked }, fieldRecord }) => ({
+    ...fieldRecord,
     type: 'checkbox',
-    initialValue: props.checked
+    initialValue: checked
   }),
   enforceProps: (props, contextProps) => ({
     checked: contextProps.get('checked')
