--- src/const/default-field-props.js
@@ -1,10 +1,0 @@
-export default {
-  type: 'text',
-  expected: true,
-  required: false,
-  disabled: false,
-  validSync: false,
-  validAsync: false,
-  validatedSync: false,
-  validatedAsync: false
-};
