+++ src/index.js
@@ -1,5 +1,4 @@
 export FormProvider from './FormProvider';
-export createField from './createField';
 export connectField from './connectField';
 export Form from './Form';
 export Condition from './Condition';
